requirements.txt
flask
tensorflow
numpy
google-cloud-storage
protobuf